requirements.txt
Flask
Flask-GraphQL
Flask-SQLAlchemy
graphene
graphene-sqlalchemy==2.3.0
psycopg2-binary
gunicorn==20.1.0